Add some more tests!
I was incorrectly handling markers. They can either have text *or* they
can have an atom index.

The sample content I used to find this also showed that sometimes Ghost
incorrectly generates a marker section and thus ruins everything. So I
put in a weird hack to try to cover for it.